fix: update blog service to handle image deletion on blog edit and delete

+++ src/services/blog.service.ts
@@ -1,5 +1,6 @@
 import Blog, { IBlog } from "../models/blog.model";
 import AppError from "../utils/appError.util";
+import { deleteImageFromFirebase } from "@/utils/firebaseStorage.util";
 
 // CREATE
 const createBlog = async (blogData: Partial<IBlog>) => {
@@ -8,11 +9,18 @@ const createBlog = async (blogData: Partial<IBlog>) => {
 
 // EDIT
 const editBlog = async (id: string, updates: Partial<IBlog>) => {
+    const currentBlog = await Blog.findById(id);
+    if (!currentBlog) throw new AppError("Blog not found", 404);
+
+    if (updates.image && currentBlog.image && currentBlog.image.startsWith("http")) {
+        await deleteImageFromFirebase(currentBlog.image);
+    }
+
     const blog = await Blog.findByIdAndUpdate(id, updates, {
         new: true,
     }).populate("author", "name username profilePicture");
 
-    return AppError.try(blog, "Blog not found", 404);
+    return blog;
 };
 
 // GET BY ID
@@ -26,8 +34,15 @@ const getBlogById = async (id: string) => {
 
 // DELETE
 const deleteBlog = async (id: string) => {
-    const blog = await Blog.findByIdAndDelete(id);
-    return AppError.try(blog, "Blog not found", 404);
+    const blog = await Blog.findById(id);
+    if (!blog) throw new AppError("Blog not found", 404);
+
+    if (blog.image && blog.image.startsWith("http")) {
+        await deleteImageFromFirebase(blog.image);
+    }
+
+    await Blog.findByIdAndDelete(id);
+    return blog;
 };
 
 // GET ALL
